+++ dnssd/instance_test.go
@@ -23,13 +23,15 @@ var _ = Describe("type ServiceInstance", func() {
 			Entry(
 				"fully populated",
 				ServiceInstance{
-					Name:        "Boardroom Printer",
-					ServiceType: "_http._tcp",
-					Domain:      "example.org",
-					TargetHost:  "boardroom-printer.example.org",
-					TargetPort:  80,
-					Priority:    10,
-					Weight:      20,
+					ServiceInstanceName: ServiceInstanceName{
+						Name:        "Boardroom Printer",
+						ServiceType: "_http._tcp",
+						Domain:      "example.org",
+					},
+					TargetHost: "boardroom-printer.example.org",
+					TargetPort: 80,
+					Priority:   10,
+					Weight:     20,
 					Attributes: []Attributes{
 						NewAttributes().
 							WithPair("txtvers", []byte{1}).
@@ -38,13 +40,15 @@ var _ = Describe("type ServiceInstance", func() {
 					TTL: 30 * time.Second,
 				},
 				ServiceInstance{
-					Name:        "Boardroom Printer",
-					ServiceType: "_http._tcp",
-					Domain:      "example.org",
-					TargetHost:  "boardroom-printer.example.org",
-					TargetPort:  80,
-					Priority:    10,
-					Weight:      20,
+					ServiceInstanceName: ServiceInstanceName{
+						Name:        "Boardroom Printer",
+						ServiceType: "_http._tcp",
+						Domain:      "example.org",
+					},
+					TargetHost: "boardroom-printer.example.org",
+					TargetPort: 80,
+					Priority:   10,
+					Weight:     20,
 					Attributes: []Attributes{
 						NewAttributes().
 							WithPair("txtvers", []byte{1}).
@@ -101,18 +105,42 @@ var _ = Describe("type ServiceInstance", func() {
 			},
 			Entry(
 				"different name",
-				ServiceInstance{Name: "Boardroom Printer"},
-				ServiceInstance{Name: "Boardroom Printer 2"},
+				ServiceInstance{
+					ServiceInstanceName: ServiceInstanceName{
+						Name: "Boardroom Printer",
+					},
+				},
+				ServiceInstance{
+					ServiceInstanceName: ServiceInstanceName{
+						Name: "Boardroom Printer 2",
+					},
+				},
 			),
 			Entry(
 				"different service type",
-				ServiceInstance{ServiceType: "_http._tcp"},
-				ServiceInstance{ServiceType: "_other._udp"},
+				ServiceInstance{
+					ServiceInstanceName: ServiceInstanceName{
+						ServiceType: "_http._tcp",
+					},
+				},
+				ServiceInstance{
+					ServiceInstanceName: ServiceInstanceName{
+						ServiceType: "_other._udp",
+					},
+				},
 			),
 			Entry(
 				"different domain",
-				ServiceInstance{Domain: "example.org"},
-				ServiceInstance{Domain: "example.com"},
+				ServiceInstance{
+					ServiceInstanceName: ServiceInstanceName{
+						Domain: "example.org",
+					},
+				},
+				ServiceInstance{
+					ServiceInstanceName: ServiceInstanceName{
+						Domain: "example.com",
+					},
+				},
 			),
 			Entry(
 				"different target host",
@@ -206,40 +234,5 @@ var _ = Describe("type ServiceInstance", func() {
 				},
 			),
 		)
-	})
-})
-
-var _ = Describe("func ServiceInstanceName()", func() {
-	It("returns the fully-qualified name, with appropriate escaping", func() {
-		d := ServiceInstanceName("Boardroom Printer.", "_http._tcp", "example.org")
-		Expect(d).To(Equal(`Boardroom\ Printer\.._http._tcp.example.org`))
-	})
-})
-
-var _ = Describe("func EscapeInstance()", func() {
-	It("escapes special characters by adding a backslash", func() {
-		n := EscapeInstance(`. '@;()"\regulartext`)
-		Expect(n).To(Equal(`\.\ \'\@\;\(\)\"\\regulartext`))
-	})
-})
-
-var _ = Describe("func ParseInstance()", func() {
-	It("unescapes dots and backslashes by removing the leading backslash", func() {
-		n, tail, err := ParseInstance(`Foo\\Bar\.`)
-		Expect(err).ShouldNot(HaveOccurred())
-		Expect(n).To(Equal(`Foo\Bar.`))
-		Expect(tail).To(BeEmpty())
-	})
-
-	It("stops after the first DNS label", func() {
-		n, tail, err := ParseInstance(`Foo\\Bar\..example.org`)
-		Expect(err).ShouldNot(HaveOccurred())
-		Expect(n).To(Equal(`Foo\Bar.`))
-		Expect(tail).To(Equal("example.org"))
-	})
-
-	It("returns an error if the name ends with the escape sequence", func() {
-		_, _, err := ParseInstance(`Foo\`)
-		Expect(err).To(MatchError("name is terminated with an escape character"))
 	})
 })